Pass arrays for model evaluation
Suppose you have a Model `m` and a set of arguments `kwargs` such that `m.eval(**kwargs)` evaluates successfully `m`. To work with a flat numpy array `x` instead of the dictionary `kwargs`, first do

    m.prepare_for_array(kwargs)

It finds the free parameters (the `None` in the defaults of the `m`) and stores the path to reach them in `_path_nones`. It also stores a deep copy of kwargs in `_template_kwargs` that will be used do define the correspondence between the flat array and dictionary representation of the arguments of `eval`.

    m.array2kwargs(x)

unpacks a flat array into a dictionary while

    m.kwargs2array(kwargs)

takes from `kwargs` the values of the free parameters of `m` and stacks them into a flat array. Finally

    m.eval_array(x)

evaluates the model based on the flat array `x`.